requirements.txt
streamlit==1.42.0
openai==1.63.2
langchain==0.3.18
load-dotenv==0.1.0
pypdf
langchain-community
tiktoken
faiss-cpu